+++ NEWS.md
@@ -1,5 +1,28 @@
 
-## First version: 1.0.0
+## Version 1.xx.xx 
+
+#### New functions
+
+`fill_symbol`: formatting function for strings. Fills a character string up to the required length. Helps to form nice messages.
+
+`set_up`: sets the argument `up` semi-globally (i.e. throughout all calls within a function).
+
+#### Change in argument names
+
+`.call_up` becomes `.up` -- retro-compatibility is ensured.
+
+#### Bug correction
+
+ . Small bug that could occur for the types "integer scalar na ok" with the argument equal to NA.
+ 
+ . Important bug when using features requiring evaluations in the environment. Now the default environment is set up appropriately.
+ 
+ . Bug when using `var(data,env)` in a `formula` class with missing `data`.
+ 
+ . Very sneaky bug when badly forming a call to \code{check_arg} to check "\code{...}" in nested functions in now caught.
+
+
+## First version: 1.0.0 (2020-04-12)
 
 This package is the outcome of over a decade of coding and developing packages in R. As a package developer I always wanted my code to be "safe": so that if a user provides arguments of the good type the functions always work, and if there's an argument of the wrong type then an informative error message is shown.